services/gateway_bff/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
python-multipart==0.0.6
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
websockets==12.0
aiofiles==23.2.1
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
email-validator==2.1.0

